+++ components/button/button.examples.js
@@ -1,6 +1,6 @@
 import React, {Component, Fragment} from 'react';
-import pencilIcon from '@jetbrains/icons/pencil.svg';
-import hourglassIcon from '@jetbrains/icons/hourglass.svg';
+import pencilIcon from '@jetbrains/icons/pencil';
+import hourglassIcon from '@jetbrains/icons/hourglass';
 
 import reactDecorator from '../../.storybook/react-decorator';