+++ components/date-picker/day.js
@@ -1,9 +1,17 @@
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import moment from 'moment';
+import addDays from 'date-fns/addDays';
+import format from 'date-fns/format';
+import getDay from 'date-fns/getDay';
+import getDate from 'date-fns/getDate';
+import isAfter from 'date-fns/isAfter';
+import isBefore from 'date-fns/isBefore';
+import isSameDay from 'date-fns/isSameDay';
+import isToday from 'date-fns/isToday';
+import startOfDay from 'date-fns/startOfDay';
 
-import {dateType, momentType, weekdays} from './consts';
+import {dateType, weekdays} from './consts';
 import styles from './date-picker.css';
 
 let hoverTO;
@@ -23,14 +31,24 @@ export default class Day extends Component {
     hoverTO = window.setTimeout(this.props.onHover, 0);
   };
 
-  isDay = date => this.props.day.isSame(date, 'day');
+  isDay = date => isSameDay(this.props.day, date);
 
   is = name => this.props[name] && this.isDay(this.props[name]);
 
-  inRange = range => range && this.props.day.isBetween(...range, 'days');
+  inRange = range => range &&
+    isAfter(startOfDay(this.props.day), startOfDay(range[0])) &&
+    isBefore(startOfDay(this.props.day), startOfDay(range[1]));
 
-  isDisabled = date => (this.props.minDate && date.isBefore(this.props.minDate, 'day')) ||
-    (this.props.maxDate && date.isAfter(this.props.maxDate, 'day'));
+  isDisabled = date => {
+    const min = this.parse(this.props.minDate);
+    const max = this.parse(this.props.maxDate);
+    return (this.props.minDate && isBefore(startOfDay(date), startOfDay(min))) ||
+      (this.props.maxDate && isAfter(startOfDay(date), startOfDay(max)));
+  };
+
+  parse(text) {
+    return this.props.parseDateInput(text);
+  }
 
   render() {
     const {
@@ -44,7 +62,7 @@ export default class Day extends Component {
     const reverse = activeRange && activeRange[1] === from;
 
     function makeSpreadRange(range) {
-      return range && [range[0], range[1].clone().add(1, 'days')];
+      return range && [range[0], addDays(range[1], 1)];
     }
 
     const spreadRange = makeSpreadRange(currentRange);
@@ -55,11 +73,11 @@ export default class Day extends Component {
         type="button"
         className={classNames(
           styles.day,
-          styles[day.format('dddd')],
+          styles[format(day, 'EEEE')],
           {
             [styles.current]: ['date', 'from', 'to'].some(this.is),
             [styles.active]: !disabled && this.is('activeDate'),
-            [styles.weekend]: [weekdays.SA, weekdays.SU].includes(day.day()),
+            [styles.weekend]: [weekdays.SA, weekdays.SU].includes(getDay(day)),
             [styles.empty]: empty,
             [styles.from]: (currentRange && this.isDay(currentRange[0]) && !reverse ||
               activeRange && this.isDay(activeRange[0])),
@@ -67,7 +85,7 @@ export default class Day extends Component {
               activeRange && this.isDay(activeRange[1]),
             [styles.between]: this.inRange(currentRange),
             [styles.activeBetween]: !disabled && this.inRange(activeRange),
-            [styles.first]: day.date() === 1,
+            [styles.first]: getDate(day) === 1,
             [styles.spread]: this.inRange(spreadRange),
             [styles.activeSpread]: !disabled && this.inRange(activeSpreadRange),
             [styles.disabled]: disabled
@@ -81,8 +99,8 @@ export default class Day extends Component {
         disabled={disabled}
       >
         {empty || (
-          <span className={classNames({[styles.today]: day.isSame(moment(), 'day')})}>
-            {day.format('D')}</span>
+          <span className={classNames({[styles.today]: isToday(day)})}>
+            {format(day, 'd')}</span>
         )}
       </button>
     );
@@ -90,12 +108,13 @@ export default class Day extends Component {
 }
 
 Day.propTypes = {
-  day: momentType,
-  from: momentType,
-  currentRange: PropTypes.arrayOf(momentType),
-  activeRange: PropTypes.arrayOf(momentType),
+  day: dateType,
+  from: dateType,
+  currentRange: PropTypes.arrayOf(dateType),
+  activeRange: PropTypes.arrayOf(dateType),
   empty: PropTypes.bool,
   onSelect: PropTypes.func,
+  parseDateInput: PropTypes.func,
   onHover: PropTypes.func,
   minDate: dateType,
   maxDate: dateType