+++ components/old-browsers-message/white-list.js
@@ -12,8 +12,7 @@ if (!SUPPORTED_BROWSERS) {
 }
 const SUPPORTED = SUPPORTED_BROWSERS || [];
 
-// TODO remove ie in 4.0
-const WHITE_LISTED_BROWSERS = ['chrome', 'firefox', 'safari', 'ie', 'edge'];
+const WHITE_LISTED_BROWSERS = ['chrome', 'firefox', 'safari', 'edge'];
 
 export const WHITE_LIST = SUPPORTED.
   reduce((acc, item) => {