+++ components/sidebar-ng/sidebar-ng.js
@@ -1,14 +1,14 @@
 import angular from 'angular';
 
-import chevronRight from '@jetbrains/icons/chevron-right.svg';
-import chevronLeft from '@jetbrains/icons/chevron-left.svg';
+import chevronRight from '@jetbrains/icons/chevron-right';
+import chevronLeft from '@jetbrains/icons/chevron-left';
 
 import PlaceUnder from '../place-under-ng/place-under-ng';
 import IconNG from '../icon-ng/icon-ng';
 import RingAngularComponent from '../global/ring-angular-component';
 import ButtonNG from '../button-ng/button-ng';
 
-import '../sidebar/sidebar.scss';
+import '../sidebar/sidebar.css';
 
 /**
  * @name Sidebar Ng