+++ components/text/text.js
@@ -1,7 +1,6 @@
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import deprecate from 'util-deprecate';
 
 import styles from './text.css';
 
@@ -9,28 +8,18 @@ import styles from './text.css';
  * @name Text
  */
 
-// TODO remove in 4.0
-const deprecateComment = deprecate(
-  () => {},
-  '<Text comment> is deprecated, use <Text info> instead'
-);
-
-
 export default class Text extends Component {
   static propTypes = {
     children: PropTypes.node,
-    comment: PropTypes.bool,
     info: PropTypes.bool,
     className: PropTypes.string
   };
 
   render() {
-    const {children, className, comment, info, ...restProps} = this.props;
-    if (comment) {
-      deprecateComment();
-    }
+    const {children, className, info, ...restProps} = this.props;
+
     const classes = classNames(styles.text, className, {
-      [styles.info]: info || comment
+      [styles.info]: info
     });
 
     return (