+++ packages/generator/app/templates/webpack.config.js
@@ -7,12 +7,6 @@ const pkgConfig = require('./package.json').config;
 
 const componentsPath = join(__dirname, pkgConfig.components);
 
-// Patch @jetbrains/ring-ui svg-sprite-loader config
-ringUiWebpackConfig.loaders.svgInlineLoader.include.push(
-  require('@jetbrains/logos'),
-  require('@jetbrains/icons')
-);
-
 const webpackConfig = () => ({
   entry: `${componentsPath}/app/app.js`,
   resolve: {
@@ -32,6 +26,12 @@ const webpackConfig = () => ({
   module: {
     rules: [
       ...ringUiWebpackConfig.config.module.rules,
+      {
+        test: /\.svg$/,
+        loader: require.resolve('svg-inline-loader'),
+        options: {removeSVGTagAttrs: false},
+        include: [require('@jetbrains/logos')]
+      },
       {
         test: /\.css$/,
         include: componentsPath,